+++ R/deployApp.R
@@ -247,7 +247,7 @@ deployApp <- function(appDir = getwd(),
       rsconnect.http.trace.json = TRUE,
       rsconnect.error.trace = TRUE
     )
-    on.exit(options(old_verbose), add = TRUE)
+    defer(options(old_verbose))
   }
 
   # install error handler if requested
@@ -258,7 +258,7 @@ deployApp <- function(appDir = getwd(),
       cat("----- Error stack trace -----\n")
       traceback(x = sys.calls(), max.lines = 3)
     })
-    on.exit(options(old_error), add = TRUE)
+    defer(options(old_error))
   }
 
   # at verbose log level, generate header
@@ -496,14 +496,22 @@ findRecordPath <- function(appDir,
   }
 }
 
-# Shinyapps defaults to public visibility.
-# Other values should be set before data is deployed.
+# Need to set _before_ deploy
 needsVisibilityChange <- function(server, application, appVisibility = NULL) {
-  if (!isCloudServer(server)) {
-    return(FALSE)
-  }
   if (is.null(appVisibility)) {
     return(FALSE)
+  }
+
+  if (isConnectServer(server)) {
+    # Defaults to private visibility
+    return(FALSE)
+  }
+
+  if (!isShinyappsServer(server)) {
+    cli::cli_abort(c(
+      "Can't change cloud app visiblity from {.fun deployApp}.",
+      i = "Please change on posit.cloud instead."
+    ))
   }
 
   cur <- application$deployment$properties$application.visibility
@@ -585,7 +593,7 @@ bundleApp <- function(appName,
       appDir = appDir,
       appFiles = appFiles,
       appPrimaryDoc = appMetadata$appPrimaryDoc)
-  on.exit(unlink(bundleDir, recursive = TRUE), add = TRUE)
+  defer(unlink(bundleDir, recursive = TRUE))
 
   # generate the manifest and write it into the bundle dir
   logger("Generate manifest.json")